+++ capon/visualization/history.py
@@ -1,6 +1,6 @@
 import plotly.express as px
 
-from capon.utils import normalize_values
+from capon.preprocessing import normalize_values
 
 
 def plot_history(stocks, anchor, title=""):
@@ -19,18 +19,37 @@ def plot_history(stocks, anchor, title=""):
 
     stocks_ts_normed = normalize_values(stocks, anchor)
 
-    fig = px.line(stocks_ts_normed, x='date', y='adjclose', color='symbol',
-                  title=f'{title}: Historical changes from {anchor}')
-    fig.update_layout(yaxis_tickformat='+%')
-    fig.update_layout(shapes=[
-        dict(
-            type='line',
-            yref='paper', y0=0, y1=1,
-            xref='x', x0=anchor, x1=anchor,
-            line=dict(width=1.5, dash='dash', color='rgba(0,0,0,0.5)'),
-        )
-    ])
+    fig = px.line(
+        stocks_ts_normed,
+        x="date",
+        y="adjclose",
+        color="symbol",
+        title=f"{title}: Historical changes from {anchor}",
+    )
+    fig.update_layout(yaxis_tickformat="+%")
+    fig.update_layout(
+        shapes=[
+            dict(
+                type="line",
+                yref="paper",
+                y0=0,
+                y1=1,
+                xref="x",
+                x0=anchor,
+                x1=anchor,
+                line=dict(width=1.5, dash="dash", color="rgba(0,0,0,0.5)"),
+            )
+        ]
+    )
     # fig.update_xaxes(rangeslider_visible=True)
     # fig.update_layout(hovermode="x")
 
     return fig
+
+
+if __name__ == "__main__":
+    import capon
+
+    stocks = capon.stock("TSLA", range="1y")
+    fig = plot_history(stocks, anchor=stocks["timestamp"].median())
+    fig.show()